+++ podcasts/Bookmarks/Profile/BookmarksProfileListController.swift
@@ -52,9 +52,8 @@ extension BookmarksProfileListController: BookmarkListRouter {
         guard let episode = bookmark.episode as? Episode else {
             return
         }
-        let controller = SharingHelper.shared.createActivityController(episode: episode, shareTime: bookmark.time)
-
-        present(controller, animated: true)
+        Analytics.track(.bookmarkShareTapped, source: viewModel.analyticsSource, properties: ["podcast_uuid": episode.podcastUuid, "episode_uuid": bookmark.episodeUuid])
+        SharingModal.show(option: .bookmark(episode, bookmark.time), from: .profile, in: self)
     }
 
     func dismissBookmarksList() {