+++ podcasts/Cancel Subscription/Cancel Subscription/Plans View/CancelSubscriptionPlanRow.swift
@@ -7,15 +7,11 @@ struct CancelSubscriptionPlanRow: View {
     var selected: Bool
     let onTap: (PlusPricingInfoModel.PlusProductPricingInfo) -> Void
 
-    var backgroundColor: Color {
-        theme.activeTheme == .indigo ? .white : theme.primaryUi01
-    }
-
     var body: some View {
         ZStack {
             Rectangle()
                 .foregroundStyle(.clear)
-                .background(backgroundColor)
+                .background(theme.primaryUi01Active)
                 .cornerRadius(8.0)
                 .frame(height: 64)
             VStack(alignment: .leading, spacing: 0) {
@@ -61,8 +57,10 @@ struct CancelSubscriptionPlanRow: View {
 extension PlusPricingInfoModel.PlusProductPricingInfo {
     fileprivate var planTitle: String {
         switch identifier {
-        case .yearly, .yearlyReferral:
+        case .yearly:
             return "Plus \(L10n.yearly.capitalized)"
+        case .yearlyReferral:
+            return "Plus \(L10n.yearly.capitalized) - Referral"
         case .monthly:
             return "Plus \(L10n.monthly.capitalized)"
         case .patronMonthly: