+++ podcasts/Common SwiftUI/Toast/ToastView.swift
@@ -109,10 +109,10 @@ struct ToastView<Style: ToastTheme>: View {
         .shadow(color: .black.opacity(0.3), radius: 10)
 
         // Animates the toast in from the bottom of the screen
+        .miniPlayerSafeAreaInset(multiplier: viewModel.aboveMiniPlayer ? 1.5 : 0)
         .offset(y: isVisible ? 0 : contentSize.height)
         .opacity(isVisible ? 1 : 0)
         .animation(ToastConstants.animation, value: isVisible)
-
         // Calculate the view size and inform the view model
         .background(GeometryReader(content: { proxy in
             Color.clear.onAppear {