+++ podcasts/Onboarding/Plus/UpgradeCard.swift
@@ -103,6 +103,8 @@ struct UpgradeCard: View {
 
     @EnvironmentObject var theme: Theme
 
+    @Environment(\.openURL) private var openURL
+
     let tier: UpgradeTier
 
     let currentPrice: Binding<PlanFrequency>
@@ -173,6 +175,17 @@ struct UpgradeCard: View {
             Text(.init("[\(purchaseTerms[safe: 3] ?? "")](\(termsOfUse))")).underline()
         }
         .foregroundColor(theme.primaryText01)
+        .environment(\.openURL, OpenURLAction { url in
+            switch url.absoluteString {
+            case privacyPolicy:
+                viewModel.privacyPolicyTapped()
+            case termsOfUse:
+                viewModel.termsOfUseTapped()
+            default:
+                break
+            }
+            return .systemAction
+        })
     }
 
     @ViewBuilder
@@ -186,7 +199,7 @@ struct UpgradeCard: View {
             viewModel.unlockTapped(.init(plan: tier.plan, frequency: currentPrice.wrappedValue))
         }, label: {
             VStack {
-                Text(tier.buttonLabel)
+                Text(purchaseTitle)
             }
             .transition(.opacity)
             .id("plus_price" + tier.title)
@@ -200,5 +213,17 @@ struct UpgradeCard: View {
                 }
             )
         }
+    }
+
+    private var purchaseTitle: String {
+        guard let subscriptionInfo = viewModel.pricingInfo(for: tier, frequency: currentPrice.wrappedValue) else {
+            return tier.buttonLabel
+        }
+
+        if subscriptionInfo.offer?.type == .freeTrial {
+            return L10n.freeTrialStartButton
+        }
+
+        return tier.buttonLabel
     }
 }