+++ podcasts/Onboarding/Plus/UpgradeLandingView.swift
@@ -101,6 +101,12 @@ struct UpgradeLandingView: View {
                                 contentIsScrollable = true
                             }
                         }
+                        .onChange(of: currentPage) { value in
+                            viewModel.changedSubscriptionTier(value)
+                        }
+                        .onChange(of: currentSubscriptionPeriod) { value in
+                            viewModel.changedSubscriptionPeriod(value)
+                        }
                     }
                 }
 
@@ -140,7 +146,7 @@ struct UpgradeLandingView: View {
         HStack(spacing: 0) {
             Spacer()
             Button(viewModel.source == .upsell ? L10n.eoyNotNow : L10n.plusSkip) {
-                viewModel.dismissTapped()
+                viewModel.dismissTapped(buttonTapped: true)
             }
             .foregroundColor(.white)
             .font(style: .body, weight: .medium)
@@ -159,7 +165,7 @@ struct UpgradeLandingView: View {
             viewModel.unlockTapped(.init(plan: selectedTier.plan, frequency: currentSubscriptionPeriod))
         }, label: {
             VStack {
-                Text(selectedTier.buttonLabel)
+                Text(purchaseTitle)
             }
             .transition(.opacity)
             .id("plus_price" + selectedTier.title)
@@ -175,6 +181,18 @@ struct UpgradeLandingView: View {
                 }
             )
         }
+    }
+
+    private var purchaseTitle: String {
+        guard let subscriptionInfo = viewModel.pricingInfo(for: selectedTier, frequency: currentSubscriptionPeriod) else {
+            return selectedTier.buttonLabel
+        }
+
+        if subscriptionInfo.offer?.type == .freeTrial {
+            return L10n.freeTrialStartButton
+        }
+
+        return selectedTier.buttonLabel
     }
 
     private var title: String {